Closes #1 and #2: append modes
Allows two modes: clearOutput() and append(). The first clears the output file before running the rule - before each test method; this is the defaut. Makes sense for the rule run one for the whole test (single test method or acting as a @ClassRule.
Append mode - appends content to an existing file. If the file doesn't exists - it's created up front